requirements.txt
tensorflow>=2.3.0,!=2.4.0,!=2.4.1,!=2.4.2
numpy
apache_beam[gcp]
PyYAML
faiss-cpu
jupyter
